---

## Background

This PR continues the work from [#6750](https://github.com/vercel/ai/pull/6750), originally targeting v4.
In v5, the `prepareStep` callback partly addressed the issue by allowing modification of `activeTools` after a tool call. However, it still does not provide a way to change the **original** tool set once the process starts.

In theory, I could send the full tool set (120 tools) to `streamText`, but this significantly impacts performance. From my measurements, a sequence of 3 tool calls takes around **14 seconds** with the full tool set, compared to **7 seconds** with this fix.

This change allows passing an updated tool set to the result of `prepareStep`, enabling more efficient tool handling without resending the entire list each time.

> **Note:** These results suggest there may be a deeper inefficiency in tool handling within the SDK. Even when only a small subset of tools is active, sending the full set still appears to slow down processing.

---

## Summary

* Added the ability to pass a new tool set to the result of `prepareStep`.
* Improves performance in cases where only a small subset of tools is required after initial calls.
* Avoids the overhead of repeatedly sending the full tool list.

---

## Manual Verification

* Tested with a scenario involving 120 tools and multiple consecutive tool calls.
* Measured execution time before and after the change:

  * **Before:** \~14 seconds for 3 consecutive calls
  * **After:** \~7 seconds for the same workload

---

## Tasks

* [ ] Add/update tests for the new behavior
* [ ] Update documentation where relevant
* [ ] Add a *patch* changeset (`pnpm changeset`)
* [ ] Fix any formatting issues (`pnpm prettier-fix`)

---

## Future Work

* Investigate the root cause of performance degradation when sending the full tool list.
* Explore further optimizations in tool management within the SDK.

---

## Related Issues

* Continues [#6750](https://github.com/vercel/ai/pull/6750)

---